+++ core/js/src/util.ts
@@ -90,3 +90,12 @@ export function mapAt<K, V>(m: Map<K, V>, k: K): V {
   }
   return ret;
 }
+
+export function _urljoin(...parts: string[]): string {
+  return parts
+    .map((x, i) =>
+      x.replace(/^\//, "").replace(i < parts.length - 1 ? /\/$/ : "", ""),
+    )
+    .filter((x) => x.trim() !== "")
+    .join("/");
+}
