+++ js/src/logger.ts
@@ -34,6 +34,7 @@ import {
   SpanObjectTypeV2,
   SpanRowIdsV2,
   gitMetadataSettingsSchema,
+  _urljoin,
 } from "@braintrust/core";
 import {
   AnyModelParam,
@@ -54,7 +55,6 @@ import {
   getCurrentUnixTimestamp,
   isEmpty,
   LazyValue,
-  _urljoin,
 } from "./util";
 import Mustache from "mustache";
 import { z } from "zod";
