--- js/src/util.test.ts
@@ -1,20 +1,0 @@
-import { test, expect } from "vitest";
-import { _urljoin } from "./util";
-
-test("_urljoin", () => {
-  expect(_urljoin("/a", "/b", "/c")).toBe("a/b/c");
-  expect(_urljoin("a", "b", "c")).toBe("a/b/c");
-  expect(_urljoin("/a/", "/b/", "/c/")).toBe("a/b/c/");
-  expect(_urljoin("a/", "b/", "c/")).toBe("a/b/c/");
-  expect(_urljoin("", "a", "b", "c")).toBe("a/b/c");
-  expect(_urljoin("a", "", "c")).toBe("a/c");
-  expect(_urljoin("/", "a", "b", "c")).toBe("a/b/c");
-  expect(_urljoin("http://example.com", "api", "v1")).toBe(
-    "http://example.com/api/v1",
-  );
-  expect(_urljoin("http://example.com/", "/api/", "/v1/")).toBe(
-    "http://example.com/api/v1/",
-  );
-  expect(_urljoin()).toBe("");
-  expect(_urljoin("a")).toBe("a");
-});
