+++ js/src/util.ts
@@ -62,12 +62,3 @@ export class LazyValue<T> {
     return this.value.hasComputed;
   }
 }
-
-export function _urljoin(...parts: string[]): string {
-  return parts
-    .map((x, i) =>
-      x.replace(/^\//, "").replace(i < parts.length - 1 ? /\/$/ : "", ""),
-    )
-    .filter((x) => x.trim() !== "")
-    .join("/");
-}
